+++ scripts/settings.js
@@ -1,6 +1,15 @@
 import { MODULE_ID } from "./module.js";
 
 export function setupSettings() {
+  game.settings.register(MODULE_ID, "disable-trigger-enable-spam", {
+    name: `${MODULE_ID}.module-settings.disable-trigger-enable-spam.name`,
+    hint: `${MODULE_ID}.module-settings.disable-trigger-enable-spam.hint`,
+    scope: "world",
+    config: true,
+    type: Boolean,
+    default: false,
+  });
+
   game.settings.register(MODULE_ID, "last-updated-settings", {
     name: "",
     hint: "",